ALG: moved game-specific files to logic sub-folder
