+++ engines/alg/logic/game_bountyhunter.h
