+++ engines/alg/logic/game_drugwars.h
