+++ engines/alg/logic/game_johnnyrock.h
