+++ engines/alg/logic/game_maddog.h
