+++ engines/alg/logic/game_maddog2.h
