+++ engines/alg/logic/game_spacepirates.h
